Moved the abstract classes to the separate folder
