+++ DimensionLogic/DimensionLoader.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Terraria.ModLoader.IO;
+using TestMod.DimensionLogic.InternalHelperClasses;
 using TestMod.Interfaces;
 
 namespace TestMod.DimensionLogic
